Get Decimal Month Values

Gets decimal month values for date/time vectors.

Usage

dtt_month_decimal(x, ...)

# S3 method for class 'Date'
dtt_month_decimal(x, ...)

# S3 method for class 'POSIXct'
dtt_month_decimal(x, ...)

Arguments

x

A date/time vector.

...

Unused.

Value

A numeric vector.

Methods (by class)

dtt_month_decimal(Date): Get numeric vector of decimal year values for a Date vector

dtt_month_decimal(POSIXct): Get numeric vector of decimal year values for a POSIXct vector

See also

Examples

x <- as.POSIXct("1990-01-03 10:00:01")
dtt_month_decimal(x)
#> [1] 1.077957