Get and Set Year Values

Gets and sets year values for date/time vectors.

dtt_years(x, ...)

dtt_years(x) <- value

dtt_set_year(x, value)

# S3 method for Date
dtt_year_decimal(x, ...)

# S3 method for POSIXct
dtt_year_decimal(x, ...)

dtt_year(x, ...)

dtt_year(x) <- value

# S3 method for Date
dtt_year(x, ...)

# S3 method for POSIXct
dtt_year(x, ...)

# S3 method for Date
dtt_year(x) <- value

# S3 method for POSIXct
dtt_year(x) <- value

Arguments

x
A date/time vector.

...
Unused.

value
A integer vector of the year value(s).

Value

An integer vector (or the modified date/time vector).

Methods (by class)

Date: Get numeric vector of decimal year values for a Date vector

POSIXct: Get numeric vector of decimal year values for a POSIXct vector

Date: Get integer vector of year values for a Date vector

POSIXct: Get integer vector of year values for a POSIXct vector

Date: Set year values for a Date vector

POSIXct: Set year values for a POSIXct vector

Examples

x <- as.Date("1990-01-02")
dtt_year(x)

#> [1] 1990

dtt_year(x) <- 11L
x

#> [1] "0011-01-02"

x <- as.POSIXct("1990-01-02 23:40:51")
dtt_year(x)

#> [1] 1990

dtt_year(x) <- 2022L
x

#> [1] "2022-01-02 23:40:51 UTC"