dttr2: Manipulate Date, POSIXct and hms Vectors

Manipulates date ('Date'), date time ('POSIXct') and time ('hms') vectors. Date/times are considered discrete and are floored whenever encountered. Times are wrapped and time zones are maintained unless explicitly altered by the user.

Manipulates date ('Date'), date time ('POSIXct') and time ('hms') vectors. Date/times are considered discrete and are floored whenever encountered. Times are wrapped and time zones are maintained unless explicitly altered by the user.

See also

Author

Maintainer: Ayla Pearson ayla@poissonconsulting.ca (ORCID)

Authors:

Joe Thorley joe@poissonconsulting.ca (ORCID)

Other contributors:

Poisson Consulting [copyright holder, funder]